+++ plugins/TimerGUIPlugin.cc
@@ -142,6 +142,10 @@ TimerGUIPlugin::TimerGUIPlugin()
   // Connect to the PreRender Gazebo signal
   this->connections.push_back(event::Events::ConnectPreRender(
                               boost::bind(&TimerGUIPlugin::PreRender, this)));
+
+  // Initialize variables
+  this->posX = 0;
+  this->posY = 0;
 }
 
 /////////////////////////////////////////////////
@@ -196,34 +200,60 @@ void TimerGUIPlugin::Load(sdf::ElementPtr _elem)
   {
     math::Vector2d p = _elem->Get<math::Vector2d>("pos");
 
-    // Check for negative x position
-    if (p.x < 0)
+    // Negative positions are counted from the ends
+    // If there are negative positions, we need to filter window resize
+    // events to reposition the timer
+    if (p.x < 0 || p.y < 0)
     {
-      gzwarn << "GUI widget x pos < 0, clamping to 0.\n";
-      p.x = 0;
+      this->parent()->installEventFilter(this);
     }
 
-    // Check for negative y position
+    if (p.x < 0)
+    {
+      if (this->parent())
+      {
+        this->posX = p.x - s.x;
+        p.x = static_cast<QWidget *>(this->parent())->width() + this->posX;
+      }
+      else
+      {
+        gzwarn << "Couldn't get parent, setting position x to zero" <<
+            std::endl;
+        p.x = 0;
+      }
+    }
+
     if (p.y < 0)
     {
-      gzwarn << "GUI widget y pos < 0, clamping to 0.\n";
-      p.y = 0;
+      if (this->parent())
+      {
+        this->posY = p.y - s.y;
+        p.y = static_cast<QWidget *>(this->parent())->height() + this->posY;
+      }
+      else
+      {
+        gzwarn << "Couldn't get parent, setting position y to zero" <<
+            std::endl;
+        p.y = 0;
+      }
     }
 
     // Check for x position greater than parent width
-    if (parent() && p.x > static_cast<QWidget*>(parent())->width())
+    if (this->parent() && p.x > static_cast<QWidget *>(this->parent())->width())
     {
       gzwarn << "GUI widget x pos > parent width, "
         << "clamping to parent width - this widget's width.\n";
-      p.x = static_cast<QWidget*>(parent())->width() - this->width();
+      p.x = static_cast<QWidget *>(this->parent())->width() - this->width();
     }
 
     // Check for y position greater than parent height
-    if (parent() && p.y > static_cast<QWidget*>(parent())->height())
+    if (this->parent() &&
+        p.y > static_cast<QWidget *>(this->parent())->height())
     {
       gzwarn << "GUI widget y pos > parent height, "
         << "clamping to parent height - this widget's height.\n";
-      p.y = static_cast<QWidget*>(parent())->height() - this->height();
+      p.y = static_cast<QWidget *>(this->parent())->height() -
+          this->height();
     }
 
     this->move(p.x, p.y);
@@ -231,8 +261,11 @@ void TimerGUIPlugin::Load(sdf::ElementPtr _elem)
   else
   {
     int xPos, yPos;
-    if (parent())
-       xPos = static_cast<QWidget*>(parent())->width() - this->width() - 10;
+    if (this->parent())
+    {
+      xPos = static_cast<QWidget *>(this->parent())->width() - this->width() -
+          10;
+    }
     else
       xPos = 600;
 
@@ -342,3 +375,27 @@ void TimerGUIPlugin::OnResetButton()
   this->Reset();
 }
 
+/////////////////////////////////////////////////
+bool TimerGUIPlugin::eventFilter(QObject *_obj, QEvent *_event)
+{
+  QWidget *widget = qobject_cast<QWidget *>(_obj);
+  if (widget == this->parent() && _event->type() == QEvent::Resize)
+  {
+    int pX = this->posX;
+    int pY = this->posY;
+
+    // Zero values mean that was a positive position, so keep the same
+    if (pX == 0)
+      pX = this->pos().x();
+    else
+      pX = widget->width() + pX;
+
+    if (pY == 0)
+      pY = this->pos().y();
+    else
+      pY = widget->height() + pY;
+
+    this->move(pX, pY);
+  }
+  return QObject::eventFilter(_obj, _event);
+}